Big restructuration of physics
I deleted lots of physics functions that were too specific to some
particuliar collision case. I made new classes and deleted methods from
the IPhysicsObject interface.

Also, ExceptionHelper is not generic in itself anymore. Only the methods
that need to be generic are. This is great, because lots of useless and
disgusting 'ExceptionHelper<int>' were killed in the process. :-)

The physics part is still under construction, though.